tests broken when empty variables
I had some problems with my Eee PC 901, that does not have a bluetooth kill
switch in /proc/ or /sys/: even if there is a modprobe-based fallback, pressing
the corresponding hotkey always left it enabled.

I finally found the problem: $BT_CTL was empty, and tested with [ -e $BT_CTL ],
that (with dash as /bin/sh) is [ -e ], that is... true! "$BT_CTL" (the empty
string) should be used instead of $BT_CTL (nothing).