Add support for ACPI button/* & video/* events; handle related model specifics.
It has been reported that, for newer kernels on some hw (1002HA, for
example), hotkey events aren't reported. On my 901, the hotkey events are
*also* reported *but* there are two keys (third and fourth silver keys) for
which only hotkey events are reported. The default is to assume that both
event types are reported; the two known keys for which this doesn't happen
are currently special-cased. (This is expected to break at some point.)

There's (so far) one button event which is reported for different keys on
different models. This is handled via the product name (sysfs, DMI).